# src/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a Web API which allows run Oryza model for Aclimate Platform",
        "version": "1.0.0",
        "title": "Aclimate Oryza Web API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "tags": [
        {
            "name": "login",
            "description": "Everything about login into into the services"
        },
        {
            "name": "run",
            "description": "Endpoint for requesting and getting oryza outputs"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [ "login" ],
                "summary": "Logs user into the system",
                "requestBody": {
                    "description": "Endpoint which allow to users authenticate into the server",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/token"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/api/v1/run": {
            "post": {
                "tags": [ "run" ],
                "summary": "Run Oryza model",
                "requestBody": {
                    "description": "Endpoint which allow to users execute oryza",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/token"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "auth": {
                "type": "object",
                "properties": {
                    "user": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "token": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "runResponse": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "format": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                }
            }
        }
    }
}